requirements.txt
PyQt5>=5.15
llama-cpp-python>=0.2.0
pyyaml>=6.0
numpy>=1.26
